+++ client/src/components/ChannelHeader/index.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import { useSelectedChannel } from '../../hooks/useSelectedChannel';
+import { useSelectedChannel } from '../../hooks';
 import { ModalController } from '../../types/modal';
 import LogoutModal from '../Modal/Logout';
 import { ChannelHeaderWrapper, ChannelHeaderLeft, ChannelHeaderRight } from './style';
