+++ client/src/components/Chat/UserConnection/UserConnection.tsx
@@ -1,8 +1,6 @@
 import React, { useEffect } from 'react';
 import useSWR from 'swr';
-import { useGroupConnection } from '../../../hooks/useGroupConnection';
-import { useSelectedGroup } from '../../../hooks/useSelectedGroup';
-import { useUserdata } from '../../../hooks/useUserdata';
+import { useGroupConnection, useSelectedGroup, useUserdata } from '../../../hooks';
 import { getFetcher } from '../../../util/fetcher';
 import { socket } from '../../../util/socket';
 import { UserConnectionWrapper, Text, UserImage, UserTile } from './style';
