+++ client/src/components/common/RestrictedRoute/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Route, RouteProps } from 'react-router-dom';
-import { useAccessControl } from '../../../hooks/useAccessControl';
+import { useAccessControl } from '../../../hooks';
 
 interface Props extends RouteProps {
   signIn: boolean;
