+++ client/src/pages/Main/index.tsx
@@ -4,9 +4,7 @@ import ChannelHeader from '../../components/ChannelHeader';
 import Chat from '../../components/Chat';
 import Meet from '../../components/Meet';
 import SideBar from '../../components/SideBar';
-import { useGroups } from '../../hooks/useGroups';
-import { useSelectedChannel } from '../../hooks/useSelectedChannel';
-import { useSelectedGroup } from '../../hooks/useSelectedGroup';
+import { useGroups, useSelectedChannel, useSelectedGroup } from '../../hooks';
 import { setSelectedChannel } from '../../redux/selectedChannel/slice';
 import { setSelectedGroup } from '../../redux/selectedGroup/slice';
 import { getURLParams } from '../../util/getURLParams';
